RF: Move export_to_figshare() out of plugins
Towards a fix for #4519
